test: move ollama api_base precedence tests to tests/test_litellm/llms/ollama/
Per @krrish-berri-2 review: tests belong under `tests/test_litellm/`, which
has a dedicated `llms/ollama/` directory with other ollama test modules.

- NEW: tests/test_litellm/llms/ollama/test_ollama_api_base_precedence.py
  (all three tests — completion, embedding, aembedding — with the same
  mock/monkeypatch setup)
- REMOVED: the three functions from tests/local_testing/test_completion.py
  (leaving the rest of that file untouched)
